Fix logical error in getPlayerChoice
The playRound function was using invalid playerSelection to return an
improper victory message. Any value other than rock, paper and scissors
was returning a victory for the player.

The current fix adds value verification to getPlayerChoice by re-prompting
the user for a new input when he inputs an invalid choice.